requirements.txt
fastapi==0.65.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
python-dotenv==0.19.0
requests==2.26.0
uvicorn==0.15.0
pydantic==1.8.2
python-decouple==3.4
pytest==6.2.5
httpx==0.19.0
aiofiles==0.7.0